Add Cherry pick function to envsetup.sh
**How it works**

-This little function is runned every time a user choose a new Target product, and verify if there is a cherry_TARGET_PRODUCT script to run and if there isn't one the process will continue without any warning or message

**Why is useful?**

It is useful because it happen that user may sometimes forget to cherry pick some essentials stuff for the compilation process and he is not aware of this only when the ROM is flashed on the device or when the compilation fails.

So by using this function the user will be warned every time he starts a new compilation to see if a cherry pick script is detected for that device.

The user just need to write a basic sh containing all the steps to cherry pick something into a repo for example build,vendor,bionic etc...
And the name to assign to the script is equivalent to this: cherry_$TARGET_PRODUCT
Which target product is your device name as for example cm_ariesve, cm_ancora.

Change-Id: I7ca8545c3227c95adca129b4bf5fc371fc0527de
Signed-off-by: nilac8991 <nilac8991@gmail.com>